# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "tatarby.shmyaks.ru API",
        "title": "TatarByHack Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/AddCompletedLesson": {
            "post": {
                "description": "Добавляет новый урок в список уроков юзера",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Добавить пройденный урок в прогресс юзера",
                "parameters": [
                    {
                        "description": "Запрос на добавление пройденного урока",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CompleteLesson"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Урок добавлен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Lesson"
                            }
                        }
                    },
                    "500": {
                        "description": "Не удалось добавить уроки",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/GetCompletedLessons": {
            "post": {
                "description": "Возвращает все пройденные юзером уроки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Получить пройденные уроки юзера",
                "responses": {
                    "200": {
                        "description": "Уроки получены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Lesson"
                            }
                        }
                    },
                    "400": {
                        "description": "Не удалось найти юзера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Не удалось получить уроки",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/GetCompletedLessonsByCourse": {
            "post": {
                "description": "Возвращает все пройденные юзером уроки в указанном курсе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Получить пройденные уроки юзера в конкретном курсе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID курса",
                        "name": "CourseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Уроки получены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Lesson"
                            }
                        }
                    },
                    "400": {
                        "description": "Не удалось найти юзера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Не удалось получить уроки",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/user_info": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Предоставляет информацию юзера.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Получить инфо пользователя",
                "responses": {
                    "200": {
                        "description": "Информация получена",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Не удалось получить информацию",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_nickname": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Изменяет никнейм пользователя и сохраняет изменения в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Изменяет никнейм пользователя",
                "parameters": [
                    {
                        "description": "New Nickname",
                        "name": "nickname",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NewNickname"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Никнейм изменен",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Не удалось изменить никнейм",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_photo": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Загружает фото пользователя и сохраняет URL в базе данных",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Загружает фото пользователя",
                "responses": {
                    "200": {
                        "description": "Фото загружено",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Не удалось загрузить фото",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.CompleteLesson": {
            "type": "object",
            "properties": {
                "courseId": {
                    "type": "integer"
                },
                "lessonId": {
                    "type": "integer"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.Lesson": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lessonName": {
                    "type": "string"
                }
            }
        },
...
        "model.NewNickname": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "Authtoken",
            "in": "cookie"
        }
    }
}